feat: refined arxiv pipeline into one script
